fix: move models into domain package
